+++ frontend/src/scenes/session-recordings/player/modal/SessionPlayerModal.tsx
@@ -2,7 +2,7 @@ import { LemonModal } from '@posthog/lemon-ui'
 import { BindLogic, useActions, useValues } from 'kea'
 import { SessionRecordingPlayer } from 'scenes/session-recordings/player/SessionRecordingPlayer'
 
-import { PlayerMeta } from '../PlayerMeta'
+import { PlayerMeta } from '../player-meta/PlayerMeta'
 import { sessionRecordingPlayerLogic, SessionRecordingPlayerLogicProps } from '../sessionRecordingPlayerLogic'
 import { sessionPlayerModalLogic } from './sessionPlayerModalLogic'
