+++ frontend/src/scenes/session-recordings/player/player-meta/PlayerMeta.scss
