+++ frontend/src/scenes/session-recordings/player/player-meta/playerMetaLogic.test.ts
@@ -1,15 +1,15 @@
 import { expectLogic } from 'kea-test-utils'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
-import { playerMetaLogic } from 'scenes/session-recordings/player/playerMetaLogic'
+import { playerMetaLogic } from 'scenes/session-recordings/player/player-meta/playerMetaLogic'
 import { sessionRecordingDataLogic } from 'scenes/session-recordings/player/sessionRecordingDataLogic'
 import { sessionRecordingPlayerLogic } from 'scenes/session-recordings/player/sessionRecordingPlayerLogic'
 
 import { useMocks } from '~/mocks/jest'
 import { initKeaTests } from '~/test/init'
 
-import recordingEventsJson from '../__mocks__/recording_events_query'
-import recordingMetaJson from '../__mocks__/recording_meta.json'
-import { snapshotsAsJSONLines } from '../__mocks__/recording_snapshots'
+import recordingEventsJson from '../../__mocks__/recording_events_query'
+import recordingMetaJson from '../../__mocks__/recording_meta.json'
+import { snapshotsAsJSONLines } from '../../__mocks__/recording_snapshots'
 
 const playerProps = { sessionRecordingId: '1', playerKey: 'playlist' }
