+++ frontend/src/scenes/session-recordings/player/sidebar/PlayerSidebarOverviewTab.tsx
@@ -2,7 +2,7 @@ import { useValues } from 'kea'
 import { PersonDisplay } from 'scenes/persons/PersonDisplay'
 import { PlayerSidebarSessionSummary } from 'scenes/session-recordings/player/sidebar/PlayerSidebarSessionSummary'
 
-import { playerMetaLogic } from '../playerMetaLogic'
+import { playerMetaLogic } from '../player-meta/playerMetaLogic'
 import { sessionRecordingPlayerLogic } from '../sessionRecordingPlayerLogic'
 import { PlayerSidebarOverviewGrid } from './PlayerSidebarOverviewGrid'
 import { PlayerSidebarOverviewOtherWatchers } from './PlayerSidebarOverviewOtherWatchers'
